refactor(tests): simplify key utilities for tests
